+++ package.json
@@ -57,6 +57,7 @@
     "mafmt": "^6.0.0",
     "multiaddr": "^5.0.0",
     "multistream-select": "^0.14.1",
+    "peer-id": "^0.11.0",
     "peer-info": "~0.14.1",
     "protons": "^1.0.1",
     "pull-abortable": "^4.1.1",
